requirements.txt
# FastAPI Backend Dependencies
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
email-validator==2.3.0

# Database
sqlalchemy==2.0.45
alembic==1.18.1

# Authentication
python-jose[cryptography]
python-dotenv==1.2.1

# Machine Learning
numpy==2.4.1
scikit-learn==1.8.0
joblib==1.5.3